add can message

+++ gripper/firmware/motor_hardware.h
@@ -23,7 +23,7 @@ void initialize_timer(motor_interrupt_callback callback);
 
 void initialize_dac();
 
-void update_pwm(uint32_t freq, uint32_t duty_cycle);
+void update_pwm(uint8_t freq, uint8_t duty_cycle);
 
 #ifdef __cplusplus
 }  // extern "C"
